requirements.txt
fastapi
uvicorn
pillow
pyzbar
streamlit
requests
ultralytics
opencv-python-headless